Support transforms in nested folders
LastGenOutput only stores the file name and does not include the relative path. The result is any transform in a nested folder will have its output wrote to the project's root folder.

The value of LastGenOutput cannot be changed in the project file to include a relative path as this causes Visual Studio to create a new file with a number appended to the end.
Visual Studio expects LastGenOutput to be in the same folder as the text template.